fix: add missing validations for endpoints

+++ src/main/java/site/soulware/cocina360/organizations/application/organization/OrganizationCommandService.java
@@ -64,6 +64,7 @@ public class OrganizationCommandService {
 
     public void handle(DeleteOrganizationCommand command) {
         Organization org = this.findOrThrow(OrganizationId.of(command.organizationId()));
+        org.requireOwner(ProfileId.of(command.requesterId()));
         this.organizationRepository.delete(org);
     }
